requirements.txt
streamlit==1.33.0
sqlalchemy==2.0.30
openai>=1.0.0
python-dotenv>=1.0.1
